Init consoleserver repository
Currently consoleserver is in the development stage, it provided
the rest api to define nodes. The nodes defined in the server will
connet to the remote and the consoleserver can work as a bridge
between remote and the client. This architecture can be used to
log the message for the sessions like `ssh`, `ipmi` or general command
etc.